agent/simple/pyproject.toml
[project]
name = "gestalt-agent-simple"
version = "0.0.1a9"
requires-python = "==3.14.*"
dependencies = [
  "gestalt-sdk==0.0.1a15",
  "anthropic==0.96.0",
  "jsonschema==4.23.0",
  "openai==2.32.0",
]

[dependency-groups]
dev = [
  "ruff==0.15.9",
  "ty==0.0.29",
  "vulture==2.16",
]

[tool.uv]
package = false
override-dependencies = [
  "typing-extensions>=4.14.1",
]

[tool.gestalt]
plugin = "provider"
agent = "provider"

[tool.gestalt.ci]
vulture-ignore-names = [
  "CreateSession",
  "GetSession",
  "ListSessions",
  "UpdateSession",
  "CreateTurn",
  "GetTurn",
  "ListTurns",
  "CancelTurn",
  "ListTurnEvents",
  "GetInteraction",
  "ListInteractions",
  "ResolveInteraction",
  "GetCapabilities",
  "configure",
  "metadata",
  "ExecuteTool",
  "do_POST",
  "log_message",
  "setUpModule",
  "tearDownModule",
  "CreateObjectStore",
  "DeleteObjectStore",
  "Get",
  "GetKey",
  "Add",
  "Put",
  "Delete",
  "Clear",
  "GetAll",
  "GetAllKeys",
  "Count",
]

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.format]
skip-magic-trailing-comma = true
